Move out ElboIntermediateVariables
