+++ src/BivariateNormals.jl
@@ -1,6 +1,88 @@
 
 module BivariateNormals
 
+using ..Types
+using ..SensitiveFloats
+import ..Util
+
+import SloanDigitalSkySurvey.WCSUtils
+
+# Types:
+export BivariateNormalDerivatives, BvnComponent, GalaxySigmaDerivs,
+       GalaxyCacheComponent
+
+# Functions:
+export eval_bvn_pdf_in_place!, get_bvn_derivs!, transform_bvn_ux_derivs!,
+       transform_bvn_derivs!, load_bvn_mixtures
+
+
+"""
+Pre-allocated memory for quantities related to derivatives of bivariate
+normals.
+"""
+type BivariateNormalDerivatives{NumType <: Number}
+
+  # Pre-allocated memory for py1, py2, and f when evaluating BVNs
+  py1::Array{NumType, 1}
+  py2::Array{NumType, 1}
+  f_pre::Array{NumType, 1}
+
+  # Derivatives of a bvn with respect to (x, sig).
+  bvn_x_d::Array{NumType, 1}
+  bvn_sig_d::Array{NumType, 1}
+  bvn_xx_h::Array{NumType, 2}
+  bvn_xsig_h::Array{NumType, 2}
+  bvn_sigsig_h::Array{NumType, 2}
+
+  # intermediate values used in d bvn / d(x, sig)
+  dpy1_dsig::Array{NumType, 1}
+  dpy2_dsig::Array{NumType, 1}
+
+  # TODO: delete this, it is now in BvnComponent
+  dsiginv_dsig::Array{NumType, 2}
+
+  # Derivatives of a bvn with respect to (u, shape)
+  bvn_u_d::Array{NumType, 1}
+  bvn_uu_h::Array{NumType, 2}
+  bvn_s_d::Array{NumType, 1}
+  bvn_ss_h::Array{NumType, 2}
+  bvn_us_h::Array{NumType, 2}
+
+  function BivariateNormalDerivatives(ThisNumType::DataType)
+    py1 = zeros(ThisNumType, 1)
+    py2 = zeros(ThisNumType, 1)
+    f_pre = zeros(ThisNumType, 1)
+
+    bvn_x_d = zeros(ThisNumType, 2)
+    bvn_sig_d = zeros(ThisNumType, 3)
+    bvn_xx_h = zeros(ThisNumType, 2, 2)
+    bvn_xsig_h = zeros(ThisNumType, 2, 3)
+    bvn_sigsig_h = zeros(ThisNumType, 3, 3)
+
+    dpy1_dsig = zeros(ThisNumType, 3)
+    dpy2_dsig = zeros(ThisNumType, 3)
+    dsiginv_dsig = zeros(ThisNumType, 3, 3)
+
+    # Derivatives wrt u.
+    bvn_u_d = zeros(ThisNumType, 2)
+    bvn_uu_h = zeros(ThisNumType, 2, 2)
+
+    # Shape deriviatives.  Here, s stands for "shape".
+    bvn_s_d = zeros(ThisNumType, length(gal_shape_ids))
+
+    # The hessians.
+    bvn_ss_h = zeros(ThisNumType, length(gal_shape_ids), length(gal_shape_ids))
+    bvn_us_h = zeros(ThisNumType, 2, length(gal_shape_ids))
+
+    new(py1, py2, f_pre,
+        bvn_x_d, bvn_sig_d, bvn_xx_h, bvn_xsig_h, bvn_sigsig_h,
+        dpy1_dsig, dpy2_dsig,
+        dsiginv_dsig,
+        bvn_u_d, bvn_uu_h, bvn_s_d, bvn_ss_h, bvn_us_h)
+  end
+end
+
+
 """
 Relevant parameters of a bivariate normal distribution.
 
@@ -52,9 +134,7 @@ immutable BvnComponent{NumType <: Number}
       else
         new{NumType}(the_mean, the_cov^-1, c * weight, zeros(0, 0))
       end
-
     end
-
 end
 
 
@@ -62,28 +142,29 @@ end
 Return quantities related to the pdf of an offset bivariate normal.
 
 Args:
+  - bvn_derivs: BivariateNormalDerivatives to be updated in place
   - bmc: A bivariate normal component
   - x: A 2x1 vector containing a mean offset to be applied to bmc
 
 Returns:
-  In elbo_vars, sets these values in place:
+  In bvn_derivs, sets these values in place:
   - py1: The first row of the precision times (x - the_mean)
   - py2: The second row of the precision times (x - the_mean)
   - The density of the bivariate normal times the weight.
 """
 function eval_bvn_pdf_in_place!{NumType <: Number}(
-    elbo_vars::ElboIntermediateVariables{NumType},
+    bvn_derivs::BivariateNormalDerivatives{NumType},
     bmc::BvnComponent{NumType}, x::Vector{Float64})
 
-  elbo_vars.bvn_derivs.py1[1] =
+  bvn_derivs.py1[1] =
     bmc.precision[1,1] * (x[1] - bmc.the_mean[1]) +
     bmc.precision[1,2] * (x[2] - bmc.the_mean[2])
-  elbo_vars.bvn_derivs.py2[1] =
+  bvn_derivs.py2[1] =
     bmc.precision[2,1] * (x[1] - bmc.the_mean[1]) +
     bmc.precision[2,2] * (x[2] - bmc.the_mean[2])
-  elbo_vars.bvn_derivs.f_pre[1] =
-    bmc.z * exp(-0.5 * ((x[1] - bmc.the_mean[1]) * elbo_vars.bvn_derivs.py1[1] +
-                        (x[2] - bmc.the_mean[2]) * elbo_vars.bvn_derivs.py2[1]))
+  bvn_derivs.f_pre[1] =
+    bmc.z * exp(-0.5 * ((x[1] - bmc.the_mean[1]) * bvn_derivs.py1[1] +
+                        (x[2] - bmc.the_mean[2]) * bvn_derivs.py2[1]))
 
   true # return type
 end
@@ -95,51 +176,15 @@ end
 """
 Calculate the value, gradient, and hessian of
   -0.5 * x' sigma^-1 x - 0.5 * log|sigma|
-with respect to x and sigma.
+with respect to x and sigma.  This assumes that
+  bvn_derivs.py1, .py2, and .f_pre have already been populated
+  witha call to eval_bvn_pdf_in_place!.
 
 Args:
   - elbo_vars: A data structure with pre-allocated intermediate variables.
   - bvn: A bivariate normal component to get derivatives for.
-  - x: The vector at which to evaluate the bvn derivs.
   - calculate_x_hess: Whether to calcualte x Hessian terms.
   - calculate_sigma_hess: Whether to calcualte sigma Hessian terms.
-"""
-function get_bvn_derivs!{NumType <: Number}(
-    elbo_vars::ElboIntermediateVariables{NumType},
-    bvn::BvnComponent{NumType},
-    x::Vector{Float64},
-    calculate_x_hess::Bool,
-    calculate_sigma_hessian::Bool)
-
-  eval_bvn_pdf_in_place!(elbo_vars, bvn, x);
-  get_bvn_derivs_in_place!(
-    elbo_vars, bvn, calculate_x_hess, calculate_sigma_hessian)
-end
-
-
-"""
-This is the function of which get_bvn_derivs!() returns the derivatives.
-It is only used for testing.
-"""
-function eval_bvn_log_density{NumType <: Number}(
-    elbo_vars::ElboIntermediateVariables{NumType},
-    bvn::BvnComponent{NumType}, x::Vector{Float64})
-
-  eval_bvn_pdf_in_place!(elbo_vars, bvn, x);
-
-  -0.5 * (
-    (x[1] - bvn.the_mean[1]) * elbo_vars.bvn_derivs.py1[1] +
-    (x[2] - bvn.the_mean[2]) * elbo_vars.bvn_derivs.py2[1] -
-    log(bvn.precision[1, 1] * bvn.precision[2, 2] - bvn.precision[1, 2] ^ 2))
-end
-
-
-"""
-Updates bvn_x_d, bvn_xx_h,
- bvn_sig_d, bvn_sigsig_h, bvn_xsig_h, in place.
- Overwrites dpy1_dsig and dpy2_dsig.
-
- Requires py1 and py2 to be set in advance.
 """
 function get_bvn_derivs!{NumType <: Number}(
     bvn_derivs::BivariateNormalDerivatives{NumType},
@@ -218,29 +263,6 @@ function get_bvn_derivs!{NumType <: Number}(
 end
 
 
-"""
-Calculate the value, gradient, and hessian of
-  -0.5 * x' sigma^-1 x - 0.5 * log|sigma|
-with respect to x and sigma.  This assumes that
-  elbo_vars..bvn_derivs.py1, .py2, and .f_pre have already been populated
-  witha call to eval_bvn_pdf_in_place!.
-
-Args:
-  - elbo_vars: A data structure with pre-allocated intermediate variables.
-  - bvn: A bivariate normal component to get derivatives for.
-  - calculate_x_hess: Whether to calcualte x Hessian terms.
-  - calculate_sigma_hess: Whether to calcualte sigma Hessian terms.
-"""
-function get_bvn_derivs_in_place!{NumType <: Number}(
-    elbo_vars::ElboIntermediateVariables{NumType},
-    bvn::BvnComponent{NumType},
-    calculate_x_hess::Bool,
-    calculate_sigma_hessian::Bool)
-
-  # TODO: get rid of redundant function when you're sure it's working
-  get_bvn_derivs!(
-    elbo_vars.bvn_derivs, bvn, calculate_x_hess, calculate_sigma_hessian)
-end
 
 
 ##############################
@@ -440,15 +462,6 @@ function transform_bvn_ux_derivs!{NumType <: Number}(
 end
 
 
-function transform_bvn_ux_derivs!{NumType <: Number}(
-    elbo_vars::ElboIntermediateVariables{NumType},
-    wcs_jacobian::Array{Float64, 2})
-
-  transform_bvn_ux_derivs!(
-    elbo_vars.bvn_derivs, wcs_jacobian, elbo_vars.calculate_hessian)
-end
-
-
 """
 Transform all the bvn derivatives from x and sigma to u and the model
 parameters.
@@ -468,8 +481,6 @@ function transform_bvn_derivs!{NumType <: Number}(
   # Transform the u derivates first.
   # bvn_x_d and bvn_xx_h should already have been set using get_bvn_derivs!()
   transform_bvn_ux_derivs!(bvn_derivs, wcs_jacobian, calculate_hessian)
-
-  # transform_bvn_derivs!(elbo_vars, gcc.bmc, wcs_jacobian)
 
   # Gradient calculations.
 
@@ -531,20 +542,6 @@ function transform_bvn_derivs!{NumType <: Number}(
 end
 
 
-"""
-Transform the bvn derivatives and hessians from (x, sigma) to the
-galaxy parameters (u, gal_shape_ids).
-
-You must have already called get_bvn_derivs!() before calling this.
-"""
-function transform_bvn_derivs!{NumType <: Number}(
-    elbo_vars::ElboIntermediateVariables{NumType},
-    gcc::GalaxyCacheComponent{NumType},
-    wcs_jacobian::Array{Float64, 2})
-
-  transform_bvn_derivs!(
-    elbo_vars.bvn_derivs, gcc.sig_sf, wcs_jacobian, elbo_vars.calculate_hessian)
-end
 
 
 """
