+++ test/test_elbo_values.jl
@@ -3,7 +3,7 @@ import Celeste: SkyImages, ModelInit
 
 using Base.Test
 using Distributions
-
+using BivariateNormals
 
 import SloanDigitalSkySurvey: SDSS, WCSUtils
